save some changes in matirx size

+++ includes/mm_mult.h
@@ -7,7 +7,7 @@ using namespace std;
 
 ////It performs M*K x K*M = M * M scheme.
 
-typedef half Data_t;
+typedef float Data_t;
 //Blocksize follows M*K x K*M = M * M scheme
 const int Block_Size_M = 4;
 const int Block_Size_K = 4;
